Added: DATA Vacuum CoP EoP 22 Oct
Added a Mathematica file for the most recent runs for EoP and CoP for the vacuum state. The first part of the file has the functions that I used to run the minimization algorithm, so it is not necessary to run it. The data points are in the second section. The parameteres are slightly different than the ones Alex used but we can see a matching behaviour for EoP.