Added: Bosonic CoP Data Hugo
Added a Mathematical File with the data obtained for bosonic Cop for a line and a circle. For a circle, the largest w calculated so far is 10, while for a line the largest w is 5. There are still calculations running for the line. I will add my own EoP calculations as well.