Modified: BCoP[\mu] Circle VS Line
Added more data and plots for BcoP[\mu] on a circle. From the recent data sets we can establish that CoP (as a function of \mu) reaches its minimum value whenever the inner product between the reference state and the target state is the biggest. That is to say, whenever the reference and target state are close to each other, complexity is minimised.